Introduce the use of function attribute (__attribute__)
Begin using it to mark functions that have a better version such as
`read_line()` which has a better version: `readLine()` that uses a
byte buffer.

Through out the code base, if we see functions that have a better
version of itself, we should definitely mark it as deprecated with
a meaningful deprecation message.

In effect, the GCC compiler will through a warning level message on
the console at compile time reporting these deprecated functions.